pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "An URL shortening service built with FastAPI and PostgreSQL"
authors = [
    {name = "Robin Bakker",email = "robin@bakeable.nl"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlalchemy[asyncio] (>=2.0.42,<3.0.0)",
    "asyncpg (>=0.29.0,<1.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "alembic (>=1.16.4,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
pytest-cov = "^6.2.1"
aiosqlite = "^0.20.0"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

